[CAS-1006] Add MentionListViewStyle and MessagePreviewStyle
