+++ stream-chat-android-ui-components/src/main/kotlin/io/getstream/chat/android/ui/mention/list/MentionListView.kt
@@ -1,7 +1,6 @@
 package io.getstream.chat.android.ui.mention.list
 
 import android.content.Context
-import android.graphics.Typeface
 import android.util.AttributeSet
 import android.widget.Toast
 import android.widget.ViewFlipper
@@ -13,14 +12,9 @@ import io.getstream.chat.android.client.models.Message
 import io.getstream.chat.android.livedata.ChatDomain
 import io.getstream.chat.android.ui.R
 import io.getstream.chat.android.ui.common.extensions.internal.createStreamThemeWrapper
-import io.getstream.chat.android.ui.common.extensions.internal.getColorCompat
-import io.getstream.chat.android.ui.common.extensions.internal.getDimension
 import io.getstream.chat.android.ui.common.extensions.internal.streamThemeInflater
-import io.getstream.chat.android.ui.common.extensions.internal.use
-import io.getstream.chat.android.ui.common.style.TextStyle
 import io.getstream.chat.android.ui.databinding.StreamUiMentionListViewBinding
 import io.getstream.chat.android.ui.mention.list.internal.MentionListAdapter
-import io.getstream.chat.android.ui.message.preview.internal.MessagePreviewView
 
 public class MentionListView : ViewFlipper {
 
@@ -52,88 +46,13 @@ public class MentionListView : ViewFlipper {
         loadMoreListener?.onLoadMoreRequested()
     }
 
+    private lateinit var style: MentionListViewStyle
+
     private fun init(attrs: AttributeSet?) {
-        parseAttrs(attrs)
-
-        context.obtainStyledAttributes(
-            attrs,
-            R.styleable.MentionListView,
-            R.attr.streamUiMentionListStyle,
-            R.style.StreamUi_MentionList
-        ).use { typedArray ->
-            typedArray.getColor(
-                R.styleable.MentionListView_streamUiBackground,
-                context.getColorCompat(R.color.stream_ui_white_snow)
-            ).let(::setBackgroundColor)
-
-            typedArray.getDrawable(
-                R.styleable.MentionListView_streamUiEmptyStateDrawable
-            ).let(binding.emptyImage::setImageDrawable)
-
-            val senderTextStyle = TextStyle.Builder(typedArray)
-                .size(
-                    R.styleable.MentionListView_streamUiSenderNameTextSize,
-                    context.getDimension(R.dimen.stream_ui_text_medium)
-                )
-                .color(
-                    R.styleable.MentionListView_streamUiSenderNameTextColor,
-                    context.getColorCompat(R.color.stream_ui_text_color_primary)
-                )
-                .font(
-                    R.styleable.MentionListView_streamUiSenderNameTextFontAssets,
-                    R.styleable.MentionListView_streamUiSenderNameTextFont
-                )
-                .style(
-                    R.styleable.MentionListView_streamUiSenderNameTextStyle,
-                    Typeface.NORMAL
-                )
-                .build()
-
-            val messageTextStyle = TextStyle.Builder(typedArray)
-                .size(
-                    R.styleable.MentionListView_streamUiMessageTextSize,
-                    context.getDimension(R.dimen.stream_ui_text_medium)
-                )
-                .color(
-                    R.styleable.MentionListView_streamUiMessageTextColor,
-                    context.getColorCompat(R.color.stream_ui_text_color_primary)
-                )
-                .font(
-                    R.styleable.MentionListView_streamUiMessageTextFontAssets,
-                    R.styleable.MentionListView_streamUiMessageTextFont
-                )
-                .style(
-                    R.styleable.MentionListView_streamUiMessageTextStyle,
-                    Typeface.NORMAL
-                )
-                .build()
-
-            val messageTimeTextStyle = TextStyle.Builder(typedArray)
-                .size(
-                    R.styleable.MentionListView_streamUiMessageTimeTextSize,
-                    context.getDimension(R.dimen.stream_ui_text_medium)
-                )
-                .color(
-                    R.styleable.MentionListView_streamUiMessageTimeTextColor,
-                    context.getColorCompat(R.color.stream_ui_text_color_primary)
-                )
-                .font(
-                    R.styleable.MentionListView_streamUiMessageTimeTextFontAssets,
-                    R.styleable.MentionListView_streamUiMessageTimeTextFont
-                )
-                .style(
-                    R.styleable.MentionListView_streamUiMessageTimeTextStyle,
-                    Typeface.NORMAL
-                )
-                .build()
-
-            val previewConfig = MessagePreviewView.PreviewStyle(
-                messageSenderTextStyle = senderTextStyle,
-                messageTextStyle = messageTextStyle,
-                messageTimeTextStyle = messageTimeTextStyle
-            )
-
-            adapter.previewStyle = previewConfig
+        style = MentionListViewStyle(context, attrs).also { style ->
+            setBackgroundColor(style.backgroundColor)
+            binding.emptyImage.setImageDrawable(style.emptyStateDrawable)
+            adapter.previewStyle = style.messagePreviewStyle
         }
 
         binding.mentionListRecyclerView.apply {
@@ -149,10 +68,6 @@ public class MentionListView : ViewFlipper {
             )
             addOnScrollListener(scrollListener)
         }
-    }
-
-    private fun parseAttrs(attrs: AttributeSet?) {
-        attrs ?: return
     }
 
     public fun showMessages(messages: List<Message>) {
