+++ stream-chat-android-ui-components/src/main/kotlin/io/getstream/chat/android/ui/message/preview/MessagePreviewStyle.kt
@@ -1,0 +1,9 @@
+package io.getstream.chat.android.ui.message.preview
+
+import io.getstream.chat.android.ui.common.style.TextStyle
+
+public data class MessagePreviewStyle(
+    val messageSenderTextStyle: TextStyle,
+    val messageTextStyle: TextStyle,
+    val messageTimeTextStyle: TextStyle,
+)
