+++ stream-chat-android-ui-components/src/main/kotlin/io/getstream/chat/android/ui/message/preview/internal/MessagePreviewView.kt
@@ -14,8 +14,8 @@ import io.getstream.chat.android.ui.common.extensions.internal.bold
 import io.getstream.chat.android.ui.common.extensions.internal.createStreamThemeWrapper
 import io.getstream.chat.android.ui.common.extensions.internal.singletonList
 import io.getstream.chat.android.ui.common.extensions.internal.streamThemeInflater
-import io.getstream.chat.android.ui.common.style.TextStyle
 import io.getstream.chat.android.ui.databinding.StreamUiMessagePreviewItemBinding
+import io.getstream.chat.android.ui.message.preview.MessagePreviewStyle
 
 internal class MessagePreviewView : FrameLayout {
 
@@ -61,8 +61,8 @@ internal class MessagePreviewView : FrameLayout {
         attrs ?: return
     }
 
-    fun styleView(previewStyle: PreviewStyle) {
-        previewStyle.run {
+    fun styleView(messagePreviewStyle: MessagePreviewStyle) {
+        messagePreviewStyle.run {
             messageSenderTextStyle.apply(binding.senderNameLabel)
             messageTextStyle.apply(binding.messageLabel)
             messageTimeTextStyle.apply(binding.messageTimeLabel)
@@ -109,10 +109,4 @@ internal class MessagePreviewView : FrameLayout {
 
         return message.text.trim()
     }
-
-    data class PreviewStyle(
-        val messageSenderTextStyle: TextStyle,
-        val messageTextStyle: TextStyle,
-        val messageTimeTextStyle: TextStyle,
-    )
 }
